refactor: standardize inputDialog in collectionsPage

+++ lib/ui/pages/collections_page.dart
@@ -2,10 +2,12 @@ import 'package:flutter/material.dart';
 import 'package:mycelium/core/stores/collection_store.dart';
 import 'package:mycelium/ui/widgets/app_bar.dart';
 import 'package:mycelium/ui/widgets/confirmation_dialog.dart';
+import 'package:mycelium/ui/widgets/input_dialog.dart';
 import 'package:mycelium/viewmodels/collections_viewmodel.dart';
 import 'package:provider/provider.dart';
 
 class CollectionsPage extends StatelessWidget {
+  const CollectionsPage({super.key});
   
   @override
   Widget build(BuildContext context) {
@@ -25,17 +27,14 @@ class CollectionsPage extends StatelessWidget {
             Padding(padding: EdgeInsetsGeometry.all(16)),
             Expanded(child: CollectionsList()),
             ElevatedButton(
-              onPressed: () {
-                showDialog(
+              onPressed: () async {
+                final name = await showInputDialog(
                   context: context,
-                  builder: (_) => CreateCollectionDialog(
-                    onSubmit: (name) async {
-                      await context
-                          .read<CollectionsViewModel>()
-                          .createCollection(name);
-                    },
-                  ),
+                  title: "Collection name",
                 );
+                if (name == null) return;
+                if (!context.mounted) return;
+                await context.read<CollectionsViewModel>().createCollection(name);
               },
               child: Text("Create collection"),
             ),
@@ -47,6 +46,8 @@ class CollectionsPage extends StatelessWidget {
 }
 
 class CollectionsList extends StatefulWidget {
+  const CollectionsList({super.key});
+  
   @override
   State<CollectionsList> createState() => _CollectionsListState();
 }
@@ -72,8 +73,8 @@ class _CollectionsListState extends State<CollectionsList> {
           return Center(
             child: ListTile(
               tileColor: collection.id == store.currentCollection?.id
-                  ? Theme.of(context).colorScheme.primary.withValues(alpha: 0.1)
-                  : null,
+              ? Theme.of(context).colorScheme.primary.withValues(alpha: 0.1)
+              : null,
               onTap: () {
                 vm.setCollection(collection.id);
               },
@@ -87,37 +88,16 @@ class _CollectionsListState extends State<CollectionsList> {
                           ListTile(
                             leading: Icon(Icons.edit),
                             title: Text("Rename"),
-                            onTap: () {
-                              showDialog(
+                            onTap: () async {
+                              final name = await showInputDialog(
                                 context: context,
-                                builder: (context) {
-                                  final controller = TextEditingController(
-                                    text: collection.name,
-                                  );
-                                  return SimpleDialog(
-                                    title: Text("New name"),
-                                    children: [
-                                      Center(
-                                        child: Padding(
-                                          padding: EdgeInsets.all(16.0),
-                                          child: TextField(
-                                            controller: controller,
-                                            autofocus: true,
-                                            onSubmitted: (value) async {
-                                              await vm.renameCollection(
-                                                collection.id,
-                                                value,
-                                              );
-                                              Navigator.pop(context);
-                                              Navigator.pop(context);
-                                            },
-                                          ),
-                                        ),
-                                      ),
-                                    ],
-                                  );
-                                },
+                                title: "New name",
+                                initialValue: collection.name
                               );
+                              if (!context.mounted) return;
+                              Navigator.pop(context);
+                              if (name == null) return;
+                              await vm.renameCollection(collection.id, name);
                             },
                           ),
                           ListTile(
@@ -131,9 +111,10 @@ class _CollectionsListState extends State<CollectionsList> {
                                 context,
                                 title: "Delete confirmation",
                                 text:
-                                    "Delete collection ${collection.name} and all data associated?",
+                                "Delete collection ${collection.name} and all data associated?",
                                 destructive: true,
                               );
+                              if (!context.mounted) return;
                               Navigator.pop(context);
                               if (confirm == true) {
                                 vm.deleteCollection(collection.id);
@@ -155,41 +136,6 @@ class _CollectionsListState extends State<CollectionsList> {
           );
         },
       ),
-    );
-  }
-}
-
-class CreateCollectionDialog extends StatefulWidget {
-  final Function(String name) onSubmit;
-
-  const CreateCollectionDialog({super.key, required this.onSubmit});
-
-  @override
-  State<CreateCollectionDialog> createState() => _CreateCollectionDialogState();
-}
-
-class _CreateCollectionDialogState extends State<CreateCollectionDialog> {
-  final TextEditingController controller = TextEditingController();
-
-  @override
-  Widget build(BuildContext context) {
-    return SimpleDialog(
-      title: Text("Collection name"),
-      children: [
-        Center(
-          child: Padding(
-            padding: EdgeInsets.all(16.0),
-            child: TextField(
-              controller: controller,
-              autofocus: true,
-              onSubmitted: (value) async {
-                await widget.onSubmit(value);
-                Navigator.pop(context);
-              },
-            ),
-          ),
-        ),
-      ],
     );
   }
 }
